commit changes when query is successful
Whenever I had a typo in a query, all my temporary tables would
disappear. This is because this kernel calls `_conn.rollback()`
whenever an error happens. This change provides the yin to that yang.
Cells that execute successfully will be followed by a commit, so that
cells that fail subsequently don’t undo more than they should.